+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.7.0 (unreleased)
+
+- Added `subvector` function
+
 ## 0.6.2 (2024-03-18)
 
 - Reduced lock contention with parallel HNSW index builds